add updates tab with basic filtering
