+++ packages/expo-dev-launcher/ios/SwiftUI/Data.swift
@@ -16,14 +16,19 @@ struct BuildInfo {
   let runtimeVersion: String
   let usesEASUpdates: Bool
   let projectUrl: String?
-  
+
   init(buildInfo: [AnyHashable: Any], updatesConfig: [AnyHashable: Any]) {
-    // Get appId and runtimeVersion from either source
     self.appId = (updatesConfig["appId"] as? String) ?? (buildInfo["appId"] as? String) ?? ""
     self.runtimeVersion = (updatesConfig["runtimeVersion"] as? String) ?? (buildInfo["runtimeVersion"] as? String) ?? ""
-    
-    // usesEASUpdates comes from updatesConfig
     self.usesEASUpdates = updatesConfig["usesEASUpdates"] as? Bool ?? false
     self.projectUrl = updatesConfig["projectUrl"] as? String
   }
+}
+
+struct RecentlyOpenedApp: Identifiable {
+  let id = UUID().uuidString
+  let name: String
+  let url: String
+  let timestamp: Date
+  let isEasUpdate: Bool?
 }
